Replace MessageInput with InputBar logic and remove duplicate
